refactor file names (working)
