+++ code/hyperdb/src/helpers/db/main.ts
@@ -1,5 +1,5 @@
 import * as ipc from './ipc.main';
 
 export { ipc };
-export { Db } from './Db.main';
+export { Db } from './Db';
 export { create } from './main.create';
